web_game_version/requirements-webgame.txt
flask
flask-cors
gunicorn
opencv-python-headless
mediapipe>=0.10.0
numpy
scikit-learn
requests
python-dotenv
